added tests for services: GeoData and WeatherCurrentApiImpl

+++ src/main/java/api/WeatherHourlyForecastApi.java
@@ -1,9 +1,10 @@
 package api;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface WeatherHourlyForecastApi
 {
-    List<Double> getTemperatureForecast(String city, int hours);
+    List<BigDecimal> getTemperatureForecast(String city, int hours);
 }
